requirements.txt

aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
beanie==1.21.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
coverage==7.2.7
discord.py==2.3.2
dnspython==2.3.0
email-validator==2.0.0.post2
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
lazy-model==0.1.0b0
motor==2.5.1
multidict==6.0.4
packaging==23.1
pluggy==1.2.0
py==1.11.0
pydantic==2.1.1
pydantic_core==2.4.0
pymongo==3.13.0
pytest==7.0.1
pytest-asyncio==0.18.3
pytest-cov==3.0.0
requests==2.31.0
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.3
yarl==1.9.2
